requirements.txt
##to install do $ 'pip install -r requitements.txt'
Flask
Flask-Login
Flask-Session
Flask_SocketIO
python-engineio
python-socketio
